Added Preliminary Sankaku Complex Support
Added support for Sankaku Complex Idol and Channel boorus. These function as expected, but cannot grab the entire list of tags if the user isn't logged in. Support for providing a user name and password is being investigated.
Updated documentation.
Bumpbed version to 1.5.0 and updated user-agent string used for web requests.